requirements.txt
streamlit>=1.28.0
google-cloud-bigquery>=3.13.0
langchain>=0.1.0
langchain-google-genai>=1.0.0
langchain-community>=0.0.13
sqlalchemy-bigquery>=1.9.0
python-dotenv>=1.0.0
db-dtypes>=1.2.0
